+++ gui/extension/CHANGELOG.md
@@ -12,6 +12,7 @@
 - BUG#37750541 Python SDK: find_unique does not require any parameter
 - BUG#38096267 Python SDK: edge case schema name is authenticate, hides method
 - BUG#38248624 Documentation is not accurate for MRS's SDK
+- BUG#37767001 Python SDK: record.delete() returns None
 -
 
 ## Changes in 1.19.15+9.4.0